refactor: move types to type file
