+++ packages/next/src/type.tsx
@@ -1,13 +1,35 @@
 import { ButtonProps } from '@alifd/next/types/button'
 import { CardProps } from '@alifd/next/types/card'
 import { RowProps, ColProps } from '@alifd/next/types/grid'
-import { IFormActions, ISchema, IEffects, IFieldError } from '@uform/types'
-
-export type TTextAlign = 'left' | 'right'
-export type TSize = 'small' | 'medium' | 'large'
-export type TLayout = 'horizontal' | 'vertical' | 'inline'
-export type TTextEl = string | JSX.Element | null
-export type TLabelAlign = 'left' | 'top' | 'inset'
+import {
+  IFormActions,
+  ISchema,
+  IEffects,
+  IFieldError,
+  Size,
+  TextAlign,
+  Layout,
+  TextEl,
+  LabelAlign
+} from '@uform/types'
+import { SwitchProps } from '@alifd/next/types/switch'
+import { GroupProps as CheckboxGroupProps } from '@alifd/next/types/checkbox'
+import { GroupProps as RadioGroupProps } from '@alifd/next/types/radio'
+import {
+  DatePickerProps,
+  RangePickerProps,
+  MonthPickerProps,
+  YearPickerProps
+} from '@alifd/next/types/date-picker'
+import { NumberPickerProps } from '@alifd/next/types/number-picker'
+import { IPasswordProps } from './fields/password'
+import { RangeProps } from '@alifd/next/types/range'
+import { RatingProps } from '@alifd/next/types/rating'
+import { InputProps, TextAreaProps } from '@alifd/next/types/input'
+import { TimePickerProps } from '@alifd/next/types/time-picker'
+import { TransferProps } from '@alifd/next/types/transfer'
+import { IUploaderProps } from './fields/upload'
+import { SelectProps } from '@alifd/next/types/select'
 
 type ColSpanType = number | string
 
@@ -16,64 +38,34 @@ export interface ColSize {
   offset?: ColSpanType
 }
 
-export interface ISchemaFormExpandProps {
-  autoAddColon?: boolean
-  className?: string
-  inline?: boolean
-  layout?: TLayout
-  maxTipsNum?: number
-  labelAlign?: TLabelAlign
-  labelTextAlign?: TTextAlign
-  labelCol?: ColSize | number
-  wrapperCol?: ColSize | number
-  size?: TSize
-  style?: React.CSSProperties
-  prefix?: string
-}
-
 export interface ILocaleMessages {
   [key: string]: string | ILocaleMessages
-}
-
-export interface ISchemaFormProps<V = unknown> {
-  actions?: IFormActions
-  initialValues?: V
-  defaultValue?: V
-  value?: V
-  editable?: boolean | ((name: string) => boolean)
-  effects?: IEffects
-  locale?: ILocaleMessages
-  schema?: ISchema
-  onChange?: (values: V) => void
-  onReset?: (values: V) => void
-  onSubmit?: (values: V) => void
-  onValidateFailed?: (fieldErrors: IFieldError[]) => void
 }
 
 export interface IFormLayoutProps {
   className?: string
   inline?: boolean
-  labelAlign?: TLabelAlign
+  labelAlign?: LabelAlign
   wrapperCol?: number
   labelCol?: number
-  labelTextAlign?: TTextAlign
-  size?: TSize
+  labelTextAlign?: TextAlign
+  size?: Size
   style?: React.CSSProperties
 }
 
 export interface IFormItemGridProps {
   cols?: Array<number | { span: number; offset: number }>
-  description?: TTextEl
+  description?: TextEl
   gutter?: number
-  title?: TTextEl
+  title?: TextEl
 }
 
 export type TFormCardOrFormBlockProps = Omit<CardProps, 'children'>
 
 export interface IFormTextBox {
   text?: string
-  title?: TTextEl
-  description?: TTextEl
+  title?: TextEl
+  description?: TextEl
   gutter?: number
 }
 
@@ -112,4 +104,57 @@ export interface IFormButtonGroupProps {
   zIndex?: number
   span?: number
   offset?: number
+}
+
+export interface SchemaFormProps<V = unknown> {
+  actions?: IFormActions
+  initialValues?: V
+  defaultValue?: V
+  value?: V
+  editable?: boolean | ((name: string) => boolean)
+  effects?: IEffects
+  locale?: ILocaleMessages
+  schema?: ISchema
+  onChange?: (values: V) => void
+  onReset?: (values: V) => void
+  onSubmit?: (values: V) => void
+  onValidateFailed?: (fieldErrors: IFieldError[]) => void
+  autoAddColon?: boolean
+  className?: string
+  inline?: boolean
+  layout?: Layout
+  maxTipsNum?: number
+  labelAlign?: LabelAlign
+  labelTextAlign?: TextAlign
+  labelCol?: ColSize | number
+  wrapperCol?: ColSize | number
+  size?: Size
+  style?: React.CSSProperties
+  prefix?: string
+}
+
+interface InternalFieldTypes {
+  boolean: SwitchProps | SelectProps
+  checkbox: CheckboxGroupProps
+  date: DatePickerProps
+  daterange: RangePickerProps
+  month: MonthPickerProps
+  // week: WeekPickerProps
+  year: YearPickerProps
+  number: NumberPickerProps | SelectProps
+  password: IPasswordProps
+  radio: RadioGroupProps
+  range: RangeProps
+  rating: RatingProps
+  string: InputProps | SelectProps
+  textarea: TextAreaProps | SelectProps
+  time: TimePickerProps
+  transfer: TransferProps
+  upload: IUploaderProps
+}
+export interface FieldProps<V, T extends string> extends ISchema<V> {
+  type?: T
+  name?: string
+  editable?: boolean
+  ['x-props']?: T extends keyof InternalFieldTypes ? InternalFieldTypes[T] : any
 }
