+++ packages/types/src/types.ts
@@ -1,6 +1,7 @@
 const isType = <T>(type: string) => (obj: unknown): obj is T =>
   obj != null && Object.prototype.toString.call(obj) === `[object ${type}]`
-export const isFn = isType<(...args: any[]) => void>('Function')
+// FIXME: isFn, isArr is incorrect
+export const isFn = isType<(...args: any[]) => any>('Function')
 export const isArr = Array.isArray || isType<unknown[]>('Array')
 export const isPlainObj = isType<object>('Object')
 export const isStr = isType<string>('String')
